+++ inst/tinytest/test_anomalize.R
@@ -18,3 +18,28 @@ expect_equal(
   , target = anm
   , info = "returns same result as `Raster*` input"
 )
+
+## errors
+expect_error(
+  anomalize(pcp, reference = "frozen_hearts")
+  , pattern = "^Expected 'reference' to inherit from .* but got"
+  , info = "throws error if 'reference' is not a raster"
+)
+
+## warnings
+ref = terra::app(pcp, fun = mean, na.rm = TRUE)
+
+expect_warning(
+  anm1 <- anomalize(
+    pcp
+    , reference = as(terra::rast(replicate(2L, ref)), "Raster")
+  )
+  , pattern = "to have a single layer, but got .* Using the first layer only"
+  , info = "throws warning if 'reference' has more than one layer"
+)
+
+expect_equal(
+  anm1
+  , target = anm
+  , info = "returns same result as single-layer 'reference'"
+)